Renamed DataObjectUtils to UID and fixed the knock-on errors.
