+++ src/main/java/uk/gov/gchq/magmacore/demo/McAssistMultInheritFromDataApp.java
@@ -14,7 +14,7 @@
 
 package uk.gov.gchq.magmacore.demo;
 
-import static uk.gov.gchq.magmacore.util.DataObjectUtils.uid;
+import static uk.gov.gchq.magmacore.util.UID.uid;
 
 import java.util.List;
